lightning: migrate tests for funding_signed transition
